add PTeamServiceDetailsSettingsView
